load services from cache in GetAll and GetByName, call discovery functions to populate all available services and their file paths
